[DOP-3714]: Rename file
