+++ src/enhanced/enhancedApp.ts
@@ -1,4 +1,4 @@
-import { listenToJobQueue } from './utils/queue/listen-to-job-queue';
+import { listenToJobQueue } from './utils/queue';
 
 async function app() {
   const job = await listenToJobQueue();
